+++ common/src/main/java/net/satisfy/wildernature/core/entity/RedWolfEntity.java
@@ -22,6 +22,7 @@ import net.minecraft.world.entity.animal.horse.Llama;
 import net.minecraft.world.entity.monster.AbstractSkeleton;
 import net.minecraft.world.entity.player.Player;
 import net.minecraft.world.item.ItemStack;
+import net.minecraft.world.item.Items;
 import net.minecraft.world.level.Level;
 import net.minecraft.world.phys.Vec3;
 import net.satisfy.wildernature.core.entity.ai.AnimationAttackGoal;
@@ -70,6 +71,15 @@ public class RedWolfEntity extends Wolf implements EntityWithAttackAnimation {
         } else {
             return SoundRegistry.RED_WOLF_AMBIENT.get();
         }
+    }
+
+    @Override
+    public boolean isFood(ItemStack stack) {
+        return stack.is(Items.BEEF)
+                || stack.is(Items.PORKCHOP)
+                || stack.is(Items.MUTTON)
+                || stack.is(Items.CHICKEN)
+                || stack.is(Items.RABBIT);
     }
 
     @Override