+++ CHANGELOG.md
@@ -1,5 +1,16 @@
 [1.1.3]
 
+**Fixed**
+- Fixed NeoForge config registration and syncing
+- Added proper breeding food items for bison, cassowary, minisheep, penguin and red wolf
+
+**Changed**
+- `TermiteSpawnEgg` Texture
+  
+***
+
+[1.1.3]
+
 **Added**
 - White deer now apply **Bad Omen** (60 min) and **Slowness II** (5 min) to the player who kills them (directly or with projectiles).